pyproject.toml
[tool.poetry]
name = "screenpear"
version = "0.1.0"
description = ""
authors = ["Dmitry Kireev <dmitry@atd.sh>"]
readme = "README.md"
packages = [
    { include = "screenpear", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
click = "^8.1.6"
opencv-python = "^4.9.0.80"
numpy = "^1.26.0"
torch = "2.2.2" # Set based on https://download.pytorch.org/whl/torch/
easyocr = "^1.7.1"
pyobjc = "^11.0"
pyobjc-framework-vision = "^11.0"
pyobjc-core = "^11.0"
pyobjc-framework-cocoa = "^11.0"
pillow = "^11.1.0"
pyobjc-framework-quartz = "^11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
